Remove the exception listener
This listener is problematic because it cannot be called along with
Symfony's default Exception Listener: they both set a
`Response`, stopping the exception propagation through other listeners.

The problem is that Symfony's default exception listener is the one
doing both the logging of the exception *and* the rendering of the
error.

The most straightforward way of getting our logs back is to simply
remove this listener.

BC breaks:

- The HTTP status code when an exception occurs is always 500. No more
  400 for `LogicException`s and 501 for `RuntimeException`s.
- The JSON response is not
  ```json
  {
    "title": "",
    "type": "/exception",
    "detail": ""
  }
  ```
  But
  ```json
  {
    "error": {
      "code":500,
      "message":"Internal Server Error"
    }
  }
  ```
  with an additional `exception` property when in dev.
- The `Content-Type` of the response is not `application/api-problem+json` but `application/json`

Enhancements:

- We have logs for exceptions
- We have less code to maintain
- Symfony's listener can handle exceptions thrown in listeners

Drawbacks:

- we loose the stack trace formatting when in dev (but we still have the
  stack trace, only in raw `json_encode($exception->getTrace())` form.

Everything concerning form errors remains the same.