Enable logging exceptions
The exception listener was not compatible with Symfony's exception
listener: they both set a Response, so they both stop the exception
event propagation.

With this change, the listener first stores a reference to the thrown
exception, then lets Symfony's exception listener log the exception and
rendering the error page (unfortunately this is done in the same
listener, an issue has been opened
[here](https://github.com/symfony/symfony/issues/18552)). Then the
listener waits for a response event, checks the status code of the
response, and if it's an error, it's returns the RestProblem custom
response as it did before, thus keeping backward compatibility.